chore(eslint-config): remove '--cache-strategy' in recommended scripts
This should speed up cache lint commands a bit, since not all files have to be read to determine if it has been changed.

The downside is that files can be modified between lint runs, but not have changed. In which case, we run all lint rules which shouldn't be neceesary.

Seems like a decent trade-off.

Signed-off-by: Dirk de Visser <github@dirkdevisser.nl>