feat(compas-convert): disable cache, final assertNotNil fixes, start with docs
Cached passes also don't execute side-effects, so we get more failures. Disabling it for now.

Had a few edge-cases in different projects for the added `assertNotNil` checks in test files, which now all run successfully through the converter.

Started with docs. It is more of an unorganized list, will make it more logical when converting a project.

Signed-off-by: Dirk de Visser <github@dirkdevisser.nl>